ui: refactor runbook webview structure and html
Refactor the RunbookWebviewPanel to align with the modular architecture
of other webviews like Workflow and LogBookmark.

1. Extract HTML generation logic into a dedicated RunbookHtmlGenerator
   class in the src/views directory.
2. Move RunbookWebviewPanel to src/views and eliminate the src/webview
   directory.
3. Extract the inline HTML template into a standalone
   resources/webview/runbook-template.html file.
4. Resolve CSP and newline character escaping issues in the new template
   to ensure shell script output renders appropriately.
5. Escape the message exit code output in the runbook webview template
   when appending it to the DOM to prevent potential Cross-Site Scripting
   (XSS) vulnerabilities.
